CP Newsroom ninjs formatter (#216)
* Updated to work for the bugs with qcodes in auto tagger.

Updated to work for the bugs with qcodes in auto tagger. Also removed the dependency to Index.js and will now use superdesk's library to fetch vocabularies.

* Update jimi_2.py

Fixed Bugs incoming on IndexCodes with non jimi tags.

* Update jimi_2.py

Fixed Issue with Categories.

* Updated to add JIMI tags in JSON.

* Update semaphore.py with updated code for scheme and qcodes.

* Updated to display proper JIMI Indexes and Categories.

* Updated to work for the bugs with qcodes in auto tagger.

Updated to work for the bugs with qcodes in auto tagger. Also removed the dependency to Index.js and will now use superdesk's library to fetch vocabularies.

* Update jimi_2.py

Fixed Bugs incoming on IndexCodes with non jimi tags.

* Updated to add JIMI tags in JSON.

* fix formatting

* Fixed the same key issue.

* Update ninjs_formatter_2.py

Removed Method defined twice issue.

* Update ninjs_formatter_2.py

Removing blank line issue

* Update jimi_2.py

Syntax issue solved

* Update ninjs_formatter_2.py

Removing White Spaces.

* Update ninjs_formatter_2.py

Removing trailing white spaces

* Update ninjs_formatter_2.py

Removing more white spaces

* Update ninjs_formatter_2.py

* Update semaphore.py

Fixed Mypy issues.

* Update semaphore.py

Removing White Spaces

* Update semaphore.py

Removed Final sets of error

* Update semaphore.py

Removing final mypy issues.

* Update semaphore.py

Fixed as per black formatter.

* Update ninjs_formatter_2.py

Black Formatted

* Update semaphore.py

Black formatted

* Update semaphore.py

Removing trailing white spaces

* Update ninjs_formatter_2.py

removing white spaces

* Update semaphore.py

Black formatting changes

* Add files via upload

* Update semaphore.py

Fixing Black Formatting

* Update scheme value in adapter.ts

* Add files via upload

Reformatted black file

* Add files via upload

Uploading Reformatted Ninjs_formatter.

* Update semaphore.py

Removing Env variables.

* Add files via upload

Black Formatted.

* Update ninjs_formatter_2.py

Updated formatter for NewsPro Output.

* Add files via upload

Black Formatting done.

* Add files via upload

* Black Formatted Formatters.

* Add files via upload

* Black Formatted

* Refactor auto-tagging component
Handle existing tags properly

* Update scheme value in adapter.ts

* use data_operation api for search and feedback

* Update and rename ninjs_newsroom_formatter.py to cp_ninjs_newsroom_formatter.py

renaming ninjs_newsroom_formatter.py to cp_ninjs_newsroom_formatter.py
changing the name of the Formatter to name = "CP Newsroom NINJS"
changing the type of the Formatter to type = "cp newsroom ninjs"

---------
